Added pull request changes
